OCPBUGS-18658: Unify agent install-complete with installer
Removed custom agent wait-for install-complete code.

Moved installer WaitForInstallComplete function from
cmd/openshift-install/main to cmd/openshift-install/command
so that the function can be made public.

Modified agent.newWaitForInstallCompleted() to use
the common WaitForInstallComplete function.

The benefit of moving agent over to the common WaitForInstallComplete
function is that the common function has a step to wait for
operators to be in a stable state before calling the cluster
installation complete.

Signed-off-by: Richard Su <rwsu@redhat.com>